modelo interfaz
