+++ src/main/java/Main.java
@@ -1,5 +1,6 @@
 import Vista.NewsGUI;
 import modelo.Modelo;
+import modelo.ModeloAPI;
 import Vista.*;
 import Controlador.*;
 import usuario.Usuario;
@@ -10,10 +11,13 @@ import java.util.concurrent.ExecutionException;
 public class Main {
     public static void main(String[] args) throws ExecutionException, InterruptedException {
         Vista vista = new NewsGUI();
-        Modelo modelo = new Modelo();
-        Controlador controlador = new Controlador(modelo, vista);
+        Modelo modeloAPI = new ModeloAPI();
+        Controlador controlador = new Controlador(modeloAPI, vista);
         Usuario usuario = new Usuario("Sergio", controlador);
 
         usuario.buscar();
     }
 }
+
+
+
