--- uml.puml
@@ -1,42 +1,0 @@
-@startuml
-class Main {
-+ {static} void main(String[])
-}
-class modelo.Modelo {
-- NewsApiClient newsApiClient
-~ ArrayList<Article> noticias
-~ String queryName
-+ CompletableFuture<List<Article>> parseEverythingToList(String)
-+ ArrayList<Article> getArticles(String)
-+ String getQueryName()
-}
-
-class Vista.NewsGUI {
-- JPanel panelPrincipal
-~ JPanel panelCabecera
-~ JLabel labelCabecera
-- JPanel panelNoticias
-- JPanel crearPanelPrincipal(String)
-- JPanel crearPanelCabecera(String)
-- JPanel crearPanelNoticias(ArrayList<Article>)
-- JPanel crearPanelTituloNoticia(Article)
-- void configurarEstiloLabelTitulo(JLabel,Article)
-~ ImageIcon getLogo()
-+ void update(ArrayList<Article>,String)
-+ void mostrarArticulos(ArrayList<Article>,String)
-}
-
-class Controlador.Controlador {
-- Vista vista
-- Modelo modelo
-}
-interface Vista.Vista {
-~ void mostrarArticulos(ArrayList<Article>,String)
-}
-
-
-Vista.Vista <|.. Vista.NewsGUI
-Vista.JFrame <|-- Vista.NewsGUI
-Controlador.Controlador .. Vista.Vista
-Controlador.Controlador .. Modelo
-@enduml
